+++ kanon/string/string_view_util.cc
@@ -6,7 +6,8 @@
 
 using namespace kanon;
 
-optional<uint64_t> kanon::StringViewToU64(StringView const &sv, int base) KANON_NOEXCEPT
+optional<uint64_t> kanon::StringViewToU64(StringView const &sv,
+                                          int base) KANON_NOEXCEPT
 {
   assert(sv.size() < 64);
   static char buf[64];
@@ -20,35 +21,38 @@ optional<uint64_t> kanon::StringViewToU64(StringView const &sv, int base) KANON_
   return ret;
 }
 
-optional<uint32_t> kanon::StringViewToU32(StringView const &sv, int base) KANON_NOEXCEPT
+optional<uint32_t> kanon::StringViewToU32(StringView const &sv,
+                                          int base) KANON_NOEXCEPT
 {
   assert(sv.size() < 32);
   static char buf[32];
   memcpy(buf, sv.data(), sv.size());
   buf[sv.size()] = 0;
   char *end = NULL;
-  uint32_t ret = strtoul(buf, &end, base);
+  uint32_t ret = (uint32_t)strtoul(buf, &end, base);
   if (ret == 0 && end == buf) {
     return make_null_optional<uint32_t>();
   }
   return ret;
 }
 
-optional<int32_t> kanon::StringViewTo32(StringView const &sv, int base) KANON_NOEXCEPT
+optional<int32_t> kanon::StringViewTo32(StringView const &sv,
+                                        int base) KANON_NOEXCEPT
 {
   assert(sv.size() < 32);
   static char buf[32];
   memcpy(buf, sv.data(), sv.size());
   buf[sv.size()] = 0;
   char *end = NULL;
-  int32_t ret = strtol(buf, &end, base);
+  int32_t ret = (int32_t)strtol(buf, &end, base);
   if (ret == 0 && end == buf) {
     return make_null_optional<int32_t>();
   }
   return ret;
 }
 
-optional<int64_t> kanon::StringViewTo64(StringView const &sv, int base) KANON_NOEXCEPT
+optional<int64_t> kanon::StringViewTo64(StringView const &sv,
+                                        int base) KANON_NOEXCEPT
 {
   assert(sv.size() < 64);
   static char buf[64];
